Inline components/TaxCenter into pages/TaxCenter
